@@ -3,8 +3,7 @@ package mobile.android.pass.settings;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.preference.PreferenceManager;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.util.Log;
 import android.view.ContextMenu;
@@ -13,7 +12,9 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.TextView;
+import android.widget.Toast;
 
+import androidx.preference.PreferenceManager;
 import mobile.android.pass.BuildConfig;
 import mobile.android.pass.R;
 import mobile.android.pass.utils.ClipboardHelper;
@@ -168,12 +169,18 @@ public class SettingsActivity extends AppCompatActivity
                 String keyToCopy = mStorageHelper.getArmoredPublicKey();
                 Log.i(TAG, "Public key: " + keyToCopy);
                 ClipboardHelper.copy(this, keyToCopy);
+                Toast.makeText(getApplicationContext(),
+                        getString(R.string.toast_copy_key), Toast.LENGTH_SHORT)
+                        .show();
                 return true;
             case R.id.action_copy_key_id:
                 Log.i(TAG, "Public key action: " + item.toString());
                 String keyID = mStorageHelper.getKeyID();
                 Log.i(TAG, "Key ID: " + keyID);
                 ClipboardHelper.copy(this, keyID);
+                Toast.makeText(getApplicationContext(),
+                        getString(R.string.toast_copy_key_id), Toast.LENGTH_SHORT)
+                        .show();
                 return true;
             case R.id.action_show_key:
                 startActivity(new Intent(this, PublicKeyActivity.class));
